Issue #2073 from the AS repository where due to the ruamel update to … (#66)
* Issue #2073 from the AS repository where due to the ruamel update to the version 0.18 A deprecated version of dump function was generating a warning when cloning a project from AS3 to the new AS4

* removing line casuing problems with CI/CD

* unit test for validation of file generated as we upgrade project from AS3 to AS4

* adjusting variable names and fixing import that broke the test

* adjusts as instructed + comparition of file making sure both backup and yaml created have the same values

* removing comparition with True

* commiting adjustments as suggested creating a more stable test and fixing comments

+++ test/unit/conftest.py
@@ -217,7 +217,8 @@ def autosubmit_config(
 
     def finalizer() -> None:
         BasicConfig.LOCAL_ROOT_DIR = original_root_dir
-        rmtree(tmp_path)
+        if tmp_path and tmp_path.exists():
+            rmtree(tmp_path)
 
     request.addfinalizer(finalizer)
